Refactoring tests and QC names, updating Readme
