+++ tests/test_analysisdriver.py
@@ -1,4 +1,3 @@
-import json
 from os.path import join, dirname
 from unittest import TestCase
 from unittest.mock import Mock
@@ -16,33 +15,8 @@ class TestAnalysisDriver(TestCase):
     assets_path = join(dirname(__file__), 'assets')
     fastq_path = join(assets_path, 'fastqs')
 
-    def __init__(self, *args, **kwargs):
-        super(TestAnalysisDriver, self).__init__(*args, **kwargs)
-        cfg.load_config_file(_etc_config('example_analysisdriver.yaml'))
 
-    @staticmethod
-    def compare_lists(observed, expected):
-        if sorted(observed) != sorted(expected):
-            print('')
-            print('observed')
-            print(observed)
-            print('expected')
-            print(expected)
-            raise AssertionError
-
-    @staticmethod
-    def query_args_from_url(url):
-        query_string = url.split('?')[1]
-        d = {}
-        for q in query_string.split('&'):
-            k, v = q.split('=')
-            if v.startswith('{') and v.endswith('}'):
-                v = json.loads(v)
-            d[k] = v
-
-        return json.loads(json.dumps(d))
-
-
+cfg.load_config_file(_etc_config('example_analysisdriver.yaml'))
 helper = TestAnalysisDriver()
 log_cfg.cfg = cfg['logging']
 log_cfg.add_stdout_handler()
