+++ tests/test_crawlers.py
@@ -1,14 +1,11 @@
 import os.path
 import json
 from unittest.mock import patch
+from egcg_core import constants as c
 
-from egcg_core.constants import ELEMENT_PROJECT_ID, ELEMENT_SAMPLE_INTERNAL_ID, ELEMENT_LIBRARY_INTERNAL_ID, \
-    ELEMENT_LANE, ELEMENT_BARCODE
-
-from tests.test_analysisdriver import TestAnalysisDriver
+from tests.test_analysisdriver import TestAnalysisDriver, NamedMock
 from analysis_driver import report_generation
 from analysis_driver.config import OutputFileConfiguration
-from tests.test_dataset import NamedMock
 
 ppath = 'analysis_driver.report_generation.'
 
@@ -65,32 +62,32 @@ class TestRunCrawler(TestCrawler):
                 'a_run_id_2_ATTACTCG': {'read_1_trimmed_bases': 284712861, 'read_2_trimmed_bases': 282625840}}
         )
         run_element1 = {
-            ELEMENT_PROJECT_ID: '10015AT',
-            ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0001',
-            ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A01',
-            ELEMENT_LANE: '1',
-            ELEMENT_BARCODE: 'ATTACTCG'
+            c.ELEMENT_PROJECT_ID: '10015AT',
+            c.ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0001',
+            c.ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A01',
+            c.ELEMENT_LANE: '1',
+            c.ELEMENT_BARCODE: 'ATTACTCG'
         }
         run_element2 = {
-            ELEMENT_PROJECT_ID: '10015AT',
-            ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0002',
-            ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A02',
-            ELEMENT_LANE: '1',
-            ELEMENT_BARCODE: 'TCCGGAGA'
+            c.ELEMENT_PROJECT_ID: '10015AT',
+            c.ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0002',
+            c.ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A02',
+            c.ELEMENT_LANE: '1',
+            c.ELEMENT_BARCODE: 'TCCGGAGA'
         }
         run_element3 = {
-            ELEMENT_PROJECT_ID: '10015AT',
-            ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0001',
-            ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A01',
-            ELEMENT_LANE: '2',
-            ELEMENT_BARCODE: 'ATTACTCG'
+            c.ELEMENT_PROJECT_ID: '10015AT',
+            c.ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0001',
+            c.ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A01',
+            c.ELEMENT_LANE: '2',
+            c.ELEMENT_BARCODE: 'ATTACTCG'
         }
         run_element4 = {
-            ELEMENT_PROJECT_ID: '10015AT',
-            ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0002',
-            ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A02',
-            ELEMENT_LANE: '2',
-            ELEMENT_BARCODE: 'TCCGGAGA'
+            c.ELEMENT_PROJECT_ID: '10015AT',
+            c.ELEMENT_SAMPLE_INTERNAL_ID: '10015AT0002',
+            c.ELEMENT_LIBRARY_INTERNAL_ID: 'LP6002014-DTP_A02',
+            c.ELEMENT_LANE: '2',
+            c.ELEMENT_BARCODE: 'TCCGGAGA'
         }
         dataset = NamedMock(
             real_name='a_run_id',
